Mimikatz Honeytoken
This script is an AutoIT logon script that launches cmd.exe as a fake
user account.  It is intended to be ran as a logon script on windows
systems.

When an attacker runs a tool such as Mimikatz against a system running
this script they will see the fake account and hopefully attempt to use
it to gain access to other machines on the network.  This enables
defenders to catch this attempt by looking for the fake account with an
IDS or by looking for failed log on attempts in Windows event logs.

Recommendation:

- Create IDS rules looking for traffic using the fake user/password.
Also consider looking for the hash values of the password that may cross
the network.
- Create log alarms looking for failed log on attempts from this user
account.