Mimikatz HoneyToken
This script is an AutoIT logon script that launches cmd.exe as a fake
user account.  It is intended to be ran as a logon script on windows
systems.

When an attacker runs a tool such as Mimikatz against a system running
this script they will see the fake account and hopefully attempt to use
it to gain access to other machines on the network.  This enables
defenders to catch this attempt by looking for the fake account with an
IDS or looking for failed log on attempts in Windows event logs.

Recommendations:

- Create IDS rules looking for traffic using the fake user and password.
Also, consider creating IDS rules looking for the NTLM hash of the fake
user.
- Create log alarms looking for failed log on attempts from this user
account.